requirements.txt
flask==3.0.0
tensorflow==2.20.0
pillow==10.1.0
numpy>=1.26.0
mysql-connector-python==9.1.0
werkzeug==3.0.0
protobuf>=5.28.0
pandas>=2.0.0
openpyxl>=3.1.0
